requirements.txt
aiofiles==0.8.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
frozenlist==1.4.1
gensim==4.3.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonify==0.5
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
packaging==23.2
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
smart-open==6.4.0
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
update-checker==0.18.0
urllib3==2.2.0
websocket-client==1.7.0
Werkzeug==3.0.1
yarl==1.9.4
